# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Keues API",
    "description": "Queue and ticket management API: locations, flows, queues, counters, tickets and devices. Authentication is done via the HttpOnly \"access_token\" cookie (JWT) set by the login and create-admin endpoints.",
    "contact": {
      "name": "Keues",
      "url": "https://www.keues.dev"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5125",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/Devices": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "Gets the registered devices, including their real-time connection status.",
        "parameters": [
          {
            "name": "LocationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "DeviceType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DeviceType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of devices.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponseOfIEnumerableOfDeviceResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponseOfIEnumerableOfDeviceResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponseOfIEnumerableOfDeviceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or business rule error.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          },
          {
            "access_token": [ ]
          }
        ]
      }
    },
    "/api/Devices/{id}": {
      "delete": {
        "tags": [
          "Devices"
        ],
        "summary": "Deletes a device from the database. Only allowed if it is disconnected.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifier of the device.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device deleted."
          },
          "400": {
            "description": "Validation or business rule error (does not exist or is still connected).",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          },
          {
            "access_token": [ ]
          }
        ]
      }
    },
...
    "/api/UserGroup": {
      "get": {
        "tags": [
          "UserGroup"
        ],
        "summary": "Get all user groups",
        "parameters": [
          {
            "name": "LocationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponseOfIEnumerableOfUserGroupBaseResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponseOfIEnumerableOfUserGroupBaseResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponseOfIEnumerableOfUserGroupBaseResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          },
          {
            "access_token": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "UserGroup"
        ],
        "summary": "Create a new user group",
        "requestBody": {
          "description": "The user group information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserGroupRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserGroupRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The created user group",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserGroupBaseResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGroupBaseResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGroupBaseResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          },
          {
            "access_token": [ ]
          }
        ]
      }
    },
    "/api/UserGroup/{id}": {
      "get": {
        "tags": [
          "UserGroup"
        ],
        "summary": "Get a user group by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the user group",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user group with the specified ID",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserGroupBaseResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGroupBaseResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGroupBaseResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          },
          {
            "access_token": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "UserGroup"
        ],
        "summary": "Update an existing user group",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the user group to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The updated user group information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserGroupRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserGroupRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated user group",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserGroupBaseResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGroupBaseResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGroupBaseResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          },
          {
            "access_token": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "UserGroup"
        ],
        "summary": "Delete a user group by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the user group to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content if the deletion is successful"
          }
        },
        "security": [
          {
            "bearer": [ ]
          },
          {
            "access_token": [ ]
          }
        ]
      }
    },
...
    "/api/Users/{id}": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update a new User. Only accessible by Admins. Returns the updated user.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          },
          {
            "access_token": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Gets a user by its identifier. Only accessible by Admins. Returns the user.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          },
          {
            "access_token": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          },
          {
            "access_token": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateUserGroupRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "userIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "DataResponseOfIEnumerableOfDeviceResponse": {
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/DeviceResponse"
            }
          },
          "pagination": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Pagination"
              }
            ]
          }
        }
      },
...
      "DataResponseOfIEnumerableOfUserGroupBaseResult": {
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/UserGroupBaseResult"
            }
          },
          "pagination": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Pagination"
              }
            ]
          }
        }
      },
...
      "DeviceResponse": {
        "required": [
          "isConnected",
          "id",
          "name",
          "type",
          "lastConnection"
        ],
        "type": "object",
        "properties": {
          "isConnected": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/DeviceType"
          },
          "lastConnection": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DeviceType": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "description": "DeviceType values: 0 = TicketMachine, 1 = Counter, 2 = Monitor.",
        "x-enum-varnames": [
          "TicketMachine",
          "Counter",
          "Monitor"
        ]
      },
...
      "ErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "GetUserResult": {
        "required": [
          "id",
          "name",
          "email",
          "locationId",
          "createdAt",
          "enabled"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "locationId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
...
      "Pagination": {
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "limit": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "UpdateUserGroupRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "userIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "UpdateUserRequest": {
        "required": [
          "name",
          "email",
          "password",
          "locationId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "UpdateUserResult": {
        "required": [
          "id",
          "name",
          "email"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "UserBasic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "UserGroupBaseResult": {
        "required": [
          "id",
          "name",
          "color",
          "locationId",
          "createdAt",
          "userIds"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "userIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserBasic"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "access_token": {
        "type": "apiKey",
        "description": "HttpOnly cookie with the JWT. It is obtained by calling POST /api/users/login or POST /api/users/create-admin.",
        "name": "access_token",
        "in": "cookie"
      },
      "bearer": {
        "type": "http",
        "description": "JWT sent in the \"Authorization: Bearer <token>\" header. Same token returned in the body by POST /api/users/login and POST /api/users/create-admin; use it for non-browser clients that cannot rely on the HttpOnly cookie.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}